+++ client/src/website/paint/index.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
-import PaintCanvas from './PaintCanvas';
+import Paint from './Paint';
 import { SocketProvider } from '../../helpers/useSocket';
 
 const container = document.querySelector('#mount-paint');
 ReactDOM.render(
     <div>
         <SocketProvider url="/paint">
-            <PaintCanvas size={896} />
+            <Paint size={896} />
         </SocketProvider>
     </div>,
-    container,
+    container
 );
